medgemma/requirements.txt
# MedGemma 1.5 4B QLoRA Fine-tuning Dependencies
# For Alzheimer's Disease Classification (CN vs AD)

# Core ML
torch>=2.0.0
transformers>=4.50.0
accelerate>=0.30.0

# QLoRA / PEFT
bitsandbytes>=0.43.0
peft>=0.10.0
trl>=0.8.0

# Data processing
datasets>=2.18.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Medical imaging
nibabel>=5.0.0
Pillow>=10.0.0
scipy>=1.10.0

# Experiment tracking
wandb>=0.16.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
